feat: add placeholder on api keys
