+++ frontend/components/playground/playground.tsx
@@ -1,21 +1,19 @@
 "use client";
 import { debounce, isEmpty } from "lodash";
-import { Loader2, PlayIcon } from "lucide-react";
 import { useParams, useRouter, useSearchParams } from "next/navigation";
 import { useCallback, useEffect, useState } from "react";
-import { Controller, FormProvider, SubmitHandler, useForm } from "react-hook-form";
+import { FormProvider, useForm } from "react-hook-form";
+import useSWR from "swr";
 
-import Messages from "@/components/playground/messages";
-import LlmSelect from "@/components/playground/messages/llm-select";
+import PlaygroundPanel from "@/components/playground/playground-panel";
+import ProvidersAlert from "@/components/playground/providers-alert";
 import { useToast } from "@/lib/hooks/use-toast";
 import { Message, Playground as PlaygroundType, PlaygroundForm } from "@/lib/playground/types";
-import { addInputs, mapMessages, parseSystemMessages, remapMessages } from "@/lib/playground/utils";
-import { streamReader } from "@/lib/utils";
+import { mapMessages, remapMessages } from "@/lib/playground/utils";
+import { ProviderApiKey } from "@/lib/settings/types";
+import { swrFetcher } from "@/lib/utils";
 
-import { Button } from "../ui/button";
-import Formatter from "../ui/formatter";
 import Header from "../ui/header";
-import { ScrollArea } from "../ui/scroll-area";
 
 const defaultMessages: Message[] = [
   {
@@ -27,13 +25,9 @@ const defaultMessages: Message[] = [
 export default function Playground({ playground }: { playground: PlaygroundType }) {
   const { replace } = useRouter();
   const params = useParams();
-  const { toast } = useToast();
   const searchParams = useSearchParams();
-  const [inputs, setInputs] = useState<string>("{}");
-  const [output, setOutput] = useState<string>("");
   const [isUpdating, setIsUpdating] = useState(false);
-
-  const [isLoading, setIsLoading] = useState(false);
+  const { toast } = useToast();
 
   const methods = useForm<PlaygroundForm>({
     defaultValues: {
@@ -42,7 +36,12 @@ export default function Playground({ playground }: { playground: PlaygroundType 
     },
   });
 
-  const { control, handleSubmit, watch, reset } = methods;
+  const { reset, watch } = methods;
+
+  const { data: apiKeys, isLoading: isApiKeysLoading } = useSWR<ProviderApiKey[]>(
+    `/api/projects/${params?.projectId}/provider-api-keys`,
+    swrFetcher
+  );
 
   const handleResetForm = async () => {
     if (playground) {
@@ -52,43 +51,6 @@ export default function Playground({ playground }: { playground: PlaygroundType 
         model: (playground.modelId as PlaygroundForm["model"]) ?? "openai:gpt-4o-mini",
         messages: isEmpty(messages) ? defaultMessages : messages,
       });
-    }
-  };
-
-  useEffect(() => {
-    handleResetForm();
-  }, []);
-
-  const submit: SubmitHandler<PlaygroundForm> = async (form) => {
-    try {
-      setIsLoading(true);
-      setOutput("");
-      const inputValues: Record<string, any> = JSON.parse(inputs);
-
-      const response = await fetch(`/api/projects/${params?.projectId}/chat`, {
-        method: "POST",
-        body: JSON.stringify({
-          projectId: params?.projectId,
-          model: form.model,
-          messages: parseSystemMessages(addInputs(form.messages, inputValues)),
-        }),
-      });
-
-      const stream = response.body?.pipeThrough(new TextDecoderStream());
-
-      if (!stream) {
-        throw new Error("No stream found.");
-      }
-
-      await streamReader(stream, (chunk) => {
-        setOutput((prev) => prev + chunk);
-      });
-    } catch (e) {
-      if (e instanceof Error) {
-        toast({ title: e.message, variant: "destructive" });
-      }
-    } finally {
-      setIsLoading(false);
     }
   };
 
@@ -113,6 +75,10 @@ export default function Playground({ playground }: { playground: PlaygroundType 
     },
     [toast]
   );
+
+  useEffect(() => {
+    handleResetForm();
+  }, []);
 
   useEffect(() => {
     if (params.playgroundId === "create" && searchParams.get("spanId")) {
@@ -142,50 +108,16 @@ export default function Playground({ playground }: { playground: PlaygroundType 
       <Header path={`playgrounds/${playground.name}`}>
         {isUpdating && <div className="w-1.5 h-1.5 rounded-full bg-yellow-400" />}
       </Header>
-      <ScrollArea className="flex-grow overflow-auto">
-        <div className="max-h-0">
-          <div className="flex flex-col gap-4 p-4">
-            <div className="flex flex-col gap-2"></div>
-            <FormProvider {...methods}>
-              <Controller
-                render={({ field: { value, onChange } }) => <LlmSelect value={value} onChange={onChange} />}
-                name="model"
-                control={control}
-              />
-              <Messages />
-            </FormProvider>
-          </div>
-          <div className="px-4">
-            <Button onClick={handleSubmit(submit)} disabled={isUpdating || isLoading}>
-              {isUpdating || isLoading ? (
-                <Loader2 className="w-4 h-4 mr-1 animate-spin" />
-              ) : (
-                <PlayIcon className="w-4 h-4 mr-1" />
-              )}
-              Run
-            </Button>
-          </div>
-          <div className="flex flex-col gap-2 p-4">
-            <div className="flex gap-4">
-              <div className="flex-1 flex flex-col gap-2">
-                <div className="text-sm font-medium">Inputs</div>
-                <Formatter
-                  value={inputs}
-                  onChange={(value) => {
-                    setInputs(value);
-                  }}
-                  editable={true}
-                  defaultMode="json"
-                />
-              </div>
-              <div className="flex-1 flex flex-col gap-2">
-                <div className="text-sm font-medium">Output</div>
-                <Formatter value={output} editable={false} defaultMode="json" />
-              </div>
-            </div>
-          </div>
+
+      {!isApiKeysLoading && !isEmpty(apiKeys) ? (
+        <FormProvider {...methods}>
+          <PlaygroundPanel apiKeys={apiKeys ?? []} isUpdating={isUpdating} />
+        </FormProvider>
+      ) : (
+        <div className="p-4">
+          <ProvidersAlert />
         </div>
-      </ScrollArea>
+      )}
     </div>
   );
 }
